Move version_certificate from announce protocol to its own file
